+++ R/bed_intersect.r
@@ -5,14 +5,22 @@
 #' in the output.
 #'
 #' @param x \code{\link{tbl_interval}}
-#' @param y \code{\link{tbl_interval}}
+#' @param ... single y  \code{\link{tbl_interval}}, multiple y \code{\link{tbl_interval}s}
+#'                   or a list of y  \code{tbl_intervals}
 #' @param invert report \code{x} intervals not in \code{y}
 #' @param suffix colname suffixes in output
-#' @param ... extra arguments (not used)
 #'
 #' @return \code{\link{tbl_interval}} with original columns from \code{x} and \code{y},
 #'   suffixed with \code{.x} and \code{.y}, and a new \code{.overlap} column
 #'   with the extent of overlap for the intersecting intervals.
+#'
+#'   If  multiple \code{y} tbls are supplied, then an additional column \code{.source} will
+#'   be reported that contains the variable names associated with each interval. All original
+#'   columns from the y tbls will be suffixed with \code{.y} in the output.
+#'   If named tbls are supplied to \code{...} (i.e \code{a = y, b = z} or
+#'    \code{list(a = y, b = z)}), then the supplied names will be reported instead
+#'   of the variable names (see examples).
+#'
 #'
 #' @template groups
 #'
@@ -56,13 +64,60 @@
 #' dplyr::mutate(res, start = pmax(start.x, start.y),
 #'                    end = pmin(end.x, end.y))
 #'
+#' z <- trbl_interval(
+#'   ~chrom, ~start, ~end, ~value,
+#'   "chr1", 150,    400,  100,
+#'   "chr1", 500,    550,  100,
+#'   "chr2", 230,    430,  200,
+#'   "chr2", 750,    900,  400
+#' )
+#'
+#' bed_intersect(x, y, z)
+#'
+#' bed_intersect(x, exons = y, introns = z)
+#'
+#' # a list of tbl_intervals can also be passed
+#' bed_intersect(x, list(exons = y, introns = z))
+#'
 #' @family multiple set operations
 #' @seealso
 #' \url{http://bedtools.readthedocs.org/en/latest/content/tools/intersect.html}
 #'
 #'
 #' @export
-bed_intersect <- function(x, y, invert = FALSE, suffix = c('.x', '.y'), ...) {
+bed_intersect <- function(x, ..., invert = FALSE, suffix = c('.x', '.y')) {
+
+  #determine if list supplied to ... or series of variables
+  if (typeof(substitute(...)) == "symbol") {
+    y_tbl <- list(...)
+    if (is.null(names(y_tbl))){
+      # name each tbl based on supplied variable
+      dots <- eval(substitute(alist(...)))
+      names(y_tbl) <- dots
+    }
+  } else {
+    # extract out just a list not a list of lists
+    y_tbl <- list(...)[[1]]
+    if (is.null(names(y_tbl))){
+      # name each tbl based on supplied variable
+      dots <- eval(substitute(alist(...)))[[1]]
+      # extract out variables from language object list(a, b, c)
+      dots <- as.character(dots)
+      dots <- dots[2:length(dots)]
+      names(y_tbl) <- as.character(dots)
+    }
+  }
+
+  multiple_tbls <- FALSE
+  if (length(y_tbl) > 1){
+    multiple_tbls <- TRUE
+    #bind_rows preserves grouping
+    y <- bind_rows(y_tbl, .id = ".source")
+    y <- select_(y, "-.source", everything(), ".source")
+  } else {
+    # only one tbl supplied, so extract out single tbl from list
+    y <- y_tbl[[1]]
+  }
 
   if (!is.tbl_interval(x)) x <- tbl_interval(x)
   if (!is.tbl_interval(y)) y <- tbl_interval(y)
@@ -77,9 +132,20 @@ bed_intersect <- function(x, y, invert = FALSE, suffix = c('.x', '.y'), ...) {
   res <- intersect_impl(x, y, suffix$x, suffix$y)
 
   if (invert) {
-    colspec <- c('chrom' = 'chrom', 'start' = 'start.x', 'end' = 'end.x')
+    colspec <- c('chrom' = 'chrom',
+                 'start' = paste0('start', suffix$x),
+                 'end' = paste0('end', suffix$x))
     res <- anti_join(x, res, by = colspec)
     res <- ungroup(res)
+  }
+
+  if (multiple_tbls) {
+    # rename .source.y to .source
+    source_col <- paste0(".source", suffix$y)
+    replace_col <- stringr::str_replace(source_col,
+                                        stringr::fixed(suffix$y), "")
+    cols <- colnames(res)
+    colnames(res) <- ifelse(cols == source_col, replace_col, cols)
   }
 
   res