+++ tests/testthat/test_intersect.r
@@ -185,3 +185,117 @@ test_that("tbls grouped by strand are processed", {
   res <- bed_intersect(group_by(x, strand), group_by(flip_strands(y), strand))
   expect_equal(nrow(res), 1)
 })
+
+test_that("invert = T, and custom suffixes don't result in failed anti_join()",{
+  x <- tibble::tribble(
+    ~chrom,   ~start,    ~end,
+    "chr3",    500,       600
+  )
+
+  y <- tibble::tribble(
+    ~chrom,   ~start,    ~end,
+    "chr1",    100,       200,
+    "chr3",    425,       475)
+
+  res <- bed_intersect(x, y, invert = T, suffix = c("a", "b"))
+  expect_equal(nrow(res), 1)
+})
+
+test_that("multiple y tbl_intervals can be passed to bed_intersect (#220)",{
+  x <- trbl_interval(
+    ~chrom, ~start, ~end,
+    "chr1", 100,    500,
+    "chr2", 200,    400,
+    "chr2", 300,    500,
+    "chr2", 800,    900
+  )
+
+  y <- trbl_interval(
+    ~chrom, ~start, ~end, ~value,
+    "chr1", 150,    400,  100,
+    "chr1", 500,    550,  100,
+    "chr2", 230,    430,  200,
+    "chr2", 350,    430,  300
+  )
+
+  z <- trbl_interval(
+    ~chrom, ~start, ~end, ~value,
+    "chr1", 150,    400,  100,
+    "chr1", 500,    550,  100,
+    "chr2", 230,    430,  200,
+    "chr2", 750,    900,  400
+  )
+
+  res <- bed_intersect(x, y, z)
+  expect_true(all(c("y", "z") %in% res$.source ))
+
+  # check that named args can be passed also
+  res <- bed_intersect(x, first_file = y, second_file = z)
+  expect_true(all(c("first_file", "second_file") %in% res$.source ))
+
+  # check that list input is parsed correctly
+  res1 <- bed_intersect(x, first_file = y, second_file = z)
+  res2 <- bed_intersect(x, list(first_file = y, second_file = z))
+  expect_equal(res1, res2)
+
+  res1 <- bed_intersect(x, y, z)
+  res2 <- bed_intersect(x, list(y, z))
+  expect_equal(res1, res2)
+
+})
+
+test_that("groups are respected when passing multiple y tbl_intervals ",{
+  x <- tibble::tribble(
+    ~chrom, ~start, ~end, ~name, ~score, ~strand,
+    "chr1", 1000,   1500, '.',   '.',     '+'
+  )
+
+  y <- tibble::tribble(
+    ~chrom, ~start, ~end, ~name, ~score, ~strand,
+    "chr1", 1000,   1200, '.',   '.',     '-'
+  )
+
+  z <- tibble::tribble(
+    ~chrom, ~start, ~end, ~name, ~score, ~strand,
+    "chr1", 1000,   1200, '.',   '.',     '+'
+  )
+  x <- group_by(x, strand)
+  y <- group_by(y, strand)
+  z <- group_by(z, strand)
+
+  res <- bed_intersect(x, y, z)
+  expect_equal(nrow(res), 1)
+})
+
+test_that("same intervals are reported with single and multiple intersection", {
+  x <- trbl_interval(
+    ~chrom, ~start, ~end,
+    "chr1", 100,    500,
+    "chr2", 200,    400,
+    "chr2", 300,    500,
+    "chr2", 800,    900
+  )
+
+  y <- trbl_interval(
+    ~chrom, ~start, ~end, ~value,
+    "chr1", 150,    400,  100,
+    "chr1", 500,    550,  100,
+    "chr2", 230,    430,  200,
+    "chr2", 350,    430,  300
+  )
+
+  z <- trbl_interval(
+    ~chrom, ~start, ~end, ~value,
+    "chr1", 150,    400,  100,
+    "chr1", 500,    550,  100,
+    "chr2", 230,    430,  200,
+    "chr2", 750,    900,  400
+  )
+  a <- bed_intersect(x, y)
+  b <- bed_intersect(x, z)
+  orig <- bind_rows(a, b) %>% arrange(chrom, start.x, start.y)
+  new <- bed_intersect(x, y, z) %>%
+    arrange(chrom, start.x, start.y) %>%
+    select(-.source)
+  expect_true(all(orig == new))
+})